e2eutil: port WaitBackupPod() and MakeBackup()
