+++ test/e2e/util.go
@@ -15,21 +15,9 @@
 package e2e
 
 import (
-	"context"
-	"fmt"
-	"net/http"
-	"time"
-
-	"github.com/coreos/etcd-operator/client/experimentalclient"
 	"github.com/coreos/etcd-operator/pkg/util/constants"
-	"github.com/coreos/etcd-operator/pkg/util/k8sutil"
-	"github.com/coreos/etcd-operator/pkg/util/retryutil"
-	"github.com/coreos/etcd-operator/test/e2e/framework"
 
 	"github.com/coreos/etcd/clientv3"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/client-go/pkg/api/v1"
 )
 
 const (
@@ -39,56 +27,6 @@ const (
 	etcdKeyFoo = "foo"
 	etcdValBar = "bar"
 )
-
-func waitBackupPodUp(f *framework.Framework, clusterName string, timeout time.Duration) error {
-	ls := labels.SelectorFromSet(map[string]string{
-		"app":          k8sutil.BackupPodSelectorAppField,
-		"etcd_cluster": clusterName,
-	}).String()
-	return retryutil.Retry(5*time.Second, int(timeout/(5*time.Second)), func() (done bool, err error) {
-		podList, err := f.KubeClient.CoreV1().Pods(f.Namespace).List(metav1.ListOptions{
-			LabelSelector: ls,
-		})
-		if err != nil {
-			return false, err
-		}
-		for i := range podList.Items {
-			if podList.Items[i].Status.Phase == v1.PodRunning {
-				return true, nil
-			}
-		}
-		return false, nil
-	})
-}
-
-func makeBackup(f *framework.Framework, clusterName string) error {
-	ls := labels.SelectorFromSet(map[string]string{
-		"app":          k8sutil.BackupPodSelectorAppField,
-		"etcd_cluster": clusterName,
-	}).String()
-	podList, err := f.KubeClient.CoreV1().Pods(f.Namespace).List(metav1.ListOptions{
-		LabelSelector: ls,
-	})
-	if err != nil {
-		return err
-	}
-	if len(podList.Items) < 1 {
-		return fmt.Errorf("no backup pod found")
-	}
-
-	// We are assuming Kubernetes pod network is accessible from test machine.
-	// TODO: remove this assumption.
-	addr := fmt.Sprintf("%s:%d", podList.Items[0].Status.PodIP, constants.DefaultBackupPodHTTPPort)
-	bc := experimentalclient.NewBackupWithAddr(&http.Client{}, "http", addr)
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer cancel()
-
-	err = bc.Request(ctx)
-	if err != nil {
-		return fmt.Errorf("backup pod (%s): %v", podList.Items[0].Name, err)
-	}
-	return nil
-}
 
 func createEtcdClient(addr string) (*clientv3.Client, error) {
 	cfg := clientv3.Config{
